+++ Editor/EditorCore/EditorMeshUtility.cs
@@ -62,7 +62,7 @@ namespace UnityEditor.ProBuilder
                 bool lightmapUVs = generateLightmapUVs || (autoLightmap && mesh.gameObject.HasStaticFlag(StaticEditorFlags.LightmapStatic));
 #endif
 
-                var usedInParticuleSystem = UnityEngine.ProBuilder.MeshUtility.IsUsedInParticleSystem(mesh);
+                var usedInParticleSystem = UnityEngine.ProBuilder.MeshUtility.IsUsedInParticleSystem(mesh);
 
                 // if generating UV2, the process is to manually split the mesh into individual triangles,
                 // generate uv2, then re-assemble with vertex collapsing where possible.
@@ -100,7 +100,7 @@ namespace UnityEditor.ProBuilder
                     UnityEngine.ProBuilder.MeshUtility.CollapseSharedVertices(umesh);
                 }
 
-                if(usedInParticuleSystem)
+                if(usedInParticleSystem)
                     UnityEngine.ProBuilder.MeshUtility.RestoreParticleSystem(mesh);
             }